+++ heylogs-api/src/main/java/internal/heylogs/spi/RuleSupport.java
@@ -9,6 +9,7 @@ import nbbrd.heylogs.Problem;
 import nbbrd.heylogs.spi.Rule;
 import nbbrd.heylogs.spi.RuleContext;
 import nbbrd.heylogs.spi.RuleIssue;
+import nbbrd.heylogs.spi.RuleSeverity;
 import nbbrd.io.text.Parser;
 
 import java.net.URL;
@@ -38,7 +39,19 @@ public final class RuleSupport {
     }
 
     private static Problem getProblemOrNull(Node node, Rule rule, RuleContext context) {
-        RuleIssue ruleIssueOrNull = rule.getRuleIssueOrNull(node, context);
-        return ruleIssueOrNull != null ? Problem.builder().rule(rule).issue(ruleIssueOrNull).build() : null;
+        RuleSeverity severity = context.findRuleSeverityOrNull(rule.getRuleId());
+        if (severity == null) severity = rule.getRuleSeverity();
+        if (!RuleSeverity.OFF.equals(severity)) {
+            RuleIssue ruleIssueOrNull = rule.getRuleIssueOrNull(node, context);
+            if (ruleIssueOrNull != null) {
+                return Problem
+                        .builder()
+                        .id(rule.getRuleId())
+                        .severity(severity)
+                        .issue(ruleIssueOrNull)
+                        .build();
+            }
+        }
+        return null;
     }
 }