+++ heylogs-api/src/main/java/nbbrd/heylogs/Config.java
@@ -1,6 +1,9 @@
 package nbbrd.heylogs;
 
+import lombok.NonNull;
 import org.jspecify.annotations.Nullable;
+
+import java.util.List;
 
 @lombok.Value
 @lombok.Builder
@@ -16,4 +19,18 @@ public class Config {
 
     @Nullable
     String forgeId;
+
+    @lombok.Singular
+    List<RuleConfig> rules;
+
+    public static final class Builder {
+
+        public @NonNull Builder versioningOf(@Nullable CharSequence versioningConfig) {
+            return versioningConfig != null ? versioning(VersioningConfig.parse(versioningConfig)) : this;
+        }
+
+        public @NonNull Builder ruleOf(@Nullable CharSequence ruleConfig) {
+            return ruleConfig != null ? rule(RuleConfig.parse(ruleConfig)) : this;
+        }
+    }
 }